// src/components/Text/types.ts
export type PossibleTag = "h1" | "h2" | "h3" | "h6" | "div" | "span" | "p";

export type TextSize =
  | "4xs"
  | "3xs"
  | "2xs"
  | "xs"
  | "sm"
  | "md"
  | "lg"
  | "xl"
  | "2xl"
  | "3xl"
  | "4xl"
  | "5xl"
  | "6xl"
  | "7xl";

export type TextWeight =
  | "thin"
  | "normal"
  | "medium"
  | "semibold"
  | "bold"
  | "extrabold"
  | "black";

export type TextTracking =
  | "tighter"
  | "tight"
  | "normal"
  | "wide"
  | "wider"
  | "widest";

export type TextLeading =
  | "loose"
  | "relaxed"
  | "normal"
  | "snug"
  | "tight"
  | "none";

export type FontVariant = "primary" | "body" | "code";